+++ hooks/android/androidCopyChannelInfo.js
@@ -2,7 +2,6 @@ const path = require('path');
 const fs = require('fs');
 const et = require('elementtree');
 const { ConfigParser } = require('cordova-common');
-const { Console } = require('console');
 
 module.exports = function (context) {
     var projectRoot = context.opts.cordova.project ? context.opts.cordova.project.root : context.opts.projectRoot;