+++ hooks/utilities.js
@@ -64,14 +64,14 @@ function getPlatformSoundPath(context, platformConfig){
   let platformPath;
 
   if(platformConfig === constants.android){
-      platformPath = path.join(projectRoot, `platforms/android/www`);
+    platformPath = path.join(projectRoot, `platforms/android/www`);
   } else {
-      let appName = getAppName(context)
-      platformPath = path.join(projectRoot, `platforms/ios/${appName}/Resources/www`);   
+    let appName = getAppName(context)
+    platformPath = path.join(projectRoot, `platforms/ios/${appName}/Resources/www`);   
   }
       
   if(!fs.existsSync(platformPath)){
-      platformPath = path.join(projectRoot, platformConfig.getWWWFolder());
+    platformPath = path.join(projectRoot, platformConfig.getWWWFolder());
   }
   
   return platformPath
@@ -83,10 +83,10 @@ function isCordovaAbove(context, version) {
   return parseInt(sp[0]) >= version;
 }
 
-
 function copyFromSourceToDestPath(defer, sourcePath, destPath) {
   fs.createReadStream(sourcePath).pipe(fs.createWriteStream(destPath))
-  .on("close", function (err) {
+  .on("close", function () {
+    console.log(`Finished copying ${sourcePath}.`);
     defer.resolve();
   })
   .on("error", function (err) {