+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
-- Chore: Update Firebase/Messaging pod to version 8.15.0. (https://outsystemsrd.atlassian.net/browse/RMET-3141)
+- Chore: Update `FirebaseMessaging` iOS pod to version `10.23.0`. This includes the Privacy Manifest (https://outsystemsrd.atlassian.net/browse/RMET-3274).
+- Chore: Update `Firebase/Messaging` iOS pod to version `8.15.0` (https://outsystemsrd.atlassian.net/browse/RMET-3141).
 
 ## [Version 2.0.1]
 - Fix: Only deal with intents from the plugin  (https://outsystemsrd.atlassian.net/browse/RMET-3013)